In this commit the registration workflow was completed, and I also deleted residual files that contained duplicate data. All data is handled by the UserDS singleton and then is the encodable object is initialised with those saved parameters. A good memory optimisation could be to create a singleton of the encodable object from the start.
There is a minor problem with the localisation view controller and the need to double click the
localisation button and the fact that the search bar is not working. I also need to add regex requirements
for the password and an email checker to be sure that the email entered is a valid email (or use a two
factor authorisation).

This issues will be flagged appropiately.